Move each class to separate file, modify init
