--- hub/codec/__init__.py
@@ -1,2 +1,0 @@
-from hub.codec.numcodecs import NumPy, Lz4, Zstd, BaseNumCodec
-from hub.codec.image_codecs import PngCodec, JpegCodec, WebPCodec, BaseImgCodec
