+++ hub/core/compression/tests/test_compression.py
@@ -1,9 +1,13 @@
 from typing import Union
 import pytest
 import numpy as np
-
-from hub.codec import BaseNumCodec, BaseImgCodec
-from hub.codec import Lz4, Zstd, NumPy, PngCodec, JpegCodec, WebPCodec
+from hub.core.compression.numpy import NUMPY
+from hub.core.compression.zstd import ZSTD
+from hub.core.compression.lz4 import LZ4
+from hub.core.compression.webp import WEBP
+from hub.core.compression.jpeg import JPEG
+from hub.core.compression.png import PNG
+from hub.core.compression import BaseNumCodec, BaseImgCodec
 
 IMG_ARRAY_SHAPES = (
     (30, 30, 3),
@@ -20,8 +24,8 @@ ARRAY_DTYPES = (
 )
 LZ4_ACCELERATIONS = (1, 50, 100)
 ZSTD_LEVELS = (1, 11, 22)
-IMG_CODECS = (PngCodec, JpegCodec, WebPCodec)
-NUM_CODECS = (Lz4, Zstd, NumPy)
+IMG_CODECS = (PNG, JPEG, WEBP)
+NUM_CODECS = (LZ4, ZSTD, NUMPY)
 
 
 def check_equals_decoded(
@@ -52,83 +56,84 @@ def check_codec_single_channel(codec: Union[BaseImgCodec]) -> None:
 @pytest.mark.parametrize("from_config", [False, True])
 @pytest.mark.parametrize("shape", IMG_ARRAY_SHAPES)
 def test_png_codec(from_config: bool, shape: tuple) -> None:
-    codec = PngCodec()
+    codec = PNG()
     if from_config:
         config = codec.get_config()
-        codec = PngCodec.from_config(config)
+        codec = codec.from_config(config)
     arr = np.ones(shape, dtype="uint8")
     check_equals_decoded(arr, codec)
 
 
 @pytest.mark.parametrize("single_channel", [False, True])
 def test_png_codec_config(single_channel: bool) -> None:
-    codec = PngCodec(single_channel=single_channel)
+    codec = PNG(single_channel=single_channel)
     check_codec_config(codec)
 
 
 @pytest.mark.parametrize("single_channel", [False, True])
 def test_png_codec_single_channel(single_channel: bool) -> None:
-    codec = PngCodec(single_channel=single_channel)
+    codec = PNG(single_channel=single_channel)
     check_codec_single_channel(codec)
 
 
 @pytest.mark.parametrize("from_config", [False, True])
 @pytest.mark.parametrize("shape", IMG_ARRAY_SHAPES)
 def test_jpeg_codec(from_config: bool, shape: tuple) -> None:
-    codec = JpegCodec()
+    compr = JPEG()
     if from_config:
-        config = codec.get_config()
-        codec = JpegCodec.from_config(config)
+        config = compr.get_config()
+        compr = JPEG.from_config(config)
     arr = np.ones(shape, dtype="uint8")
-    check_equals_decoded(arr, codec)
+    check_equals_decoded(arr, compr)
 
 
 @pytest.mark.parametrize("single_channel", [False, True])
 def test_jpeg_codec_config(single_channel: bool) -> None:
-    codec = JpegCodec(single_channel=single_channel)
+    jpeg_compressor = compression.__dict__["JPEG"]
+    codec = jpeg_compressor(single_channel=single_channel)
     check_codec_config(codec)
 
 
 @pytest.mark.parametrize("single_channel", [False, True])
 def test_jpeg_codec_single_channel(single_channel: bool) -> None:
-    codec = JpegCodec(single_channel=single_channel)
+    codec = JPEG(single_channel=single_channel)
     check_codec_single_channel(codec)
 
 
 @pytest.mark.parametrize("from_config", [False, True])
 @pytest.mark.parametrize("shape", IMG_ARRAY_SHAPES)
 def test_webp_codec(from_config: bool, shape: tuple) -> None:
-    codec = WebPCodec()
+    codec = WEBP()
     if from_config:
         config = codec.get_config()
-        codec = WebPCodec.from_config(config)
+        codec = WEBP.from_config(config)
     arr = np.ones(shape, dtype="uint8")
     check_equals_decoded(arr, codec)
 
 
 @pytest.mark.parametrize("single_channel", [False, True])
 def test_webp_codec_config(single_channel: bool) -> None:
-    codec = WebPCodec(single_channel=single_channel)
+    codec = WEBP(single_channel=single_channel)
     check_codec_config(codec)
 
 
 @pytest.mark.parametrize("single_channel", [False, True])
 def test_webp_codec_single_channel(single_channel: bool) -> None:
-    codec = WebPCodec(single_channel=single_channel)
+    codec = WEBP(single_channel=single_channel)
     check_codec_single_channel(codec)
 
 
 @pytest.mark.parametrize("acceleration", LZ4_ACCELERATIONS)
 @pytest.mark.parametrize("shape", GENERIC_ARRAY_SHAPES)
 def test_lz4(acceleration: int, shape: tuple) -> None:
-    codec = Lz4(acceleration=acceleration)
+    codec = LZ4(acceleration=acceleration)
     arr = np.random.rand(*shape)
     check_equals_decoded(arr, codec)
 
 
 @pytest.mark.parametrize("shape", GENERIC_ARRAY_SHAPES)
 def test_numpy(shape: tuple) -> None:
-    codec = NumPy()
+    codec = NUMPY()
     arr = np.random.rand(*shape)
     check_equals_decoded(arr, codec)
 
@@ -136,7 +141,7 @@ def test_numpy(shape: tuple) -> None:
 @pytest.mark.parametrize("level", ZSTD_LEVELS)
 @pytest.mark.parametrize("shape", GENERIC_ARRAY_SHAPES)
 def test_zstd(level: int, shape: tuple) -> None:
-    codec = Zstd(level=level)
+    codec = ZSTD(level=level)
     arr = np.random.rand(*shape)
     check_equals_decoded(arr, codec)
 
@@ -153,22 +158,22 @@ def test_base_name():
 
 def test_kwargs():
     with pytest.raises(ValueError):
-        compressor = Lz4(another_kwarg=1)
+        compressor = LZ4(another_kwarg=1)
     with pytest.raises(ValueError):
-        compressor = Zstd(another_kwarg=1)
+        compressor = ZSTD(another_kwarg=1)
     with pytest.raises(ValueError):
-        compressor = PngCodec(another_kwarg=1)
+        compressor = PNG(another_kwarg=1)
     with pytest.raises(ValueError):
-        compressor = JpegCodec(another_kwarg=1)
+        compressor = JPEG(another_kwarg=1)
     with pytest.raises(ValueError):
-        compressor = WebPCodec(another_kwarg=1)
+        compressor = WEBP(another_kwarg=1)
 
 
 def get_compr_input(compressor, shape):
     comp_input = np.random.randint(0, 255, size=shape, dtype="uint8")
     if compressor in NUM_CODECS:
         comp_input = comp_input.tobytes()
-    if compressor == WebPCodec:
+    if compressor == WEBP:
         compressor = compressor(quality=85)
     else:
         compressor = compressor()
